renamed literals to constants
